Split Razer Chroma types into separate files
