+++ Assets/_Project/Scripts/RazerChroma.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 2e5ab907ea4b42958223c267afa07614
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
