+++ Assets/_Project/Scripts/RazerChroma/ChromaEffectSettings.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 076dae0a0de943d1a90edfe4deb9ad60
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
