+++ Assets/_Project/Scripts/RazerChroma/RazerChromaController.cs
@@ -7,35 +7,6 @@ using ChromaSDK;
 
 public sealed class RazerChromaController : MonoBehaviour
 {
-    public enum RazerStage
-    {
-        Base,
-        ComboLow,
-        ComboMid,
-        ComboHigh,
-        Rainbow,
-        RainbowPlus,
-    }
-
-    public enum RazerEffectType
-    {
-        Solid,
-        PulseTwoColor,
-        FlowTwoColor,
-        Rainbow,
-    }
-
-    [Serializable]
-    public sealed class ChromaEffectSettings
-    {
-        public RazerEffectType effectType = RazerEffectType.PulseTwoColor;
-        public Color colorA = new(0.10f, 0.20f, 0.25f, 1f);
-        public Color colorB = new(0.08f, 0.12f, 0.18f, 1f);
-        [Range(0f, 1f)] public float intensity = 0.35f;
-        [Min(0f)] public float speed = 0.12f;
-        [Range(0f, 1f)] public float rainbowSaturation = 0.7f;
-    }
-
     [Serializable]
     public sealed class ComboLayerSettings
     {
