+++ Assets/_Project/Scripts/RazerChroma/RazerEffectType.cs
@@ -1,0 +1,7 @@
+public enum RazerEffectType
+{
+    Solid,
+    PulseTwoColor,
+    FlowTwoColor,
+    Rainbow,
+}
