+++ Assets/_Project/Scripts/RazerChroma/RazerEffectType.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 5a0b887c623c4cda80d5db45c39180c0
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
