+++ Assets/_Project/Scripts/RazerChroma/RazerStage.cs
@@ -1,0 +1,9 @@
+public enum RazerStage
+{
+    Base,
+    ComboLow,
+    ComboMid,
+    ComboHigh,
+    Rainbow,
+    RainbowPlus,
+}
