+++ Assets/_Project/Scripts/RazerChroma/RazerStage.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: dd9284eafe6f43378cb682fef8cbe64f
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
